Refactored the whole runner code.
This allows to run multiple bots in the same runner, and also allows workers to
process not only updates, but everything (will be useful later for timers).

I've tried to remove as much pickle as I can from the queues, avoiding to pass
Bot or TelegramAPI objects around. This was done because pickle likes to
complain loudly when it receives some objects it don't like. Hopefully this
improves the problem reported in #2.