Adapt botogram to the recent API changes.
The GroupChat object was removed, and a generic Chat object was introduced
instead of the old one. This object is used for private conversations, group
chats and channels, so the Message.chat attribute won't contain anymore the
User or the GroupChat object, but only the Chat one.

Also the API change introduced support for managing channels. Details of the
implementation are tracked on issue #7.