Deprecate pass_* decorators and automatically fill hooks arguments.
Before this commit, you had to request all the arguments an hook provides in
the function declaration. If you also needed the bot instance or the shared
memory, you needed to use two different decorators to get those.

That was really verbose, so this commit removes all the above things: now you
can request from the function declaration only the things you need, including
the bot instance and the shared memory, in the order you want:

   @bot.command("test")
   def test_command(chat, args, shared):
       pass

The decorators @botogram.pass_bot and @botogram.pass_shared were also
been deprecated, and they will be removed in the first release.