Nice exceptions when the IPC server crashes
Before this commit, when the IPC server crashes an endless stream of
BrokenPipeError is printed to stdout, which makes nearly impossible to
debug what's happening (an exception on the IPC server).

This commit hides those exceptions, and tells the user to kill the
runner in order to gain control of the terminal again.