Add locks support in shared memory
Before this commit there was no simple way to deal with concurrency
issues in botogram: threading locks don't work in a multiprocessing
environment, and multiprocessing locks are just hard to implement right
in the runner.

Because of that this commit introduces locks support in the shared
memory. They're implemented with threading locks out of the runner, and
with the IPC server in the runner. They also have the same API as the
threading ones.

  with shared.lock("my-lock"):
      # ...
      pass

  lock = shared.lock("another-lock")
  lock.acquire()
  # ...
  lock.release()