Fix a bunch of errors in the Windows support
Before this commit, when you use any hook except the command one on
Windows, your bot crashes because pickle doesn't know how to serialize
the hook. That was because botogram changed the actual function of the
hook, and pickle don't like that.

This commit refactors almost all the hooks creation logic, moving away
from modifying the user's function on the fly.