Refactored the Bot._call internal method
Before this commit, a bug considered all the hooks as part of the main
component, for shared memory. This caused serious issues with shared
memory initializers, since they didn't work for external components.

Now the Bot._call method accepts the component ID of the hook, and
provides shared memory as an argument only if the ID is provided.

This is a regression introduced in 350a78a.