Fix syntax detection screwed up by wrong URLs (fixes #28)
Before this commit, if you had two underscores in an URL the whole message was
marked as markdown. This caused issues because Telegram clients refuses to mark
links as links if they contain syntax in them.

This commit strips URLs and email addresses from the messages before checking
if they're markdown or not, so this bug is fixed. Also, there are new tests to
prevent this issue from happening in the future.