Fix strip_urls not stripping domains with dashes (fixes #32)
Before this commit, an url with dashes in the domain was not stripped
correctly by the botogram.utils.strip_urls functions, which caused
issues in the syntax detector (recognizing URLs as Markdown).

This commit fixes the regular expression used to strip URLs, and also
adds some unit tests for that function in order to prevent issues like
this one in the future.