Add support for fetching an user's avatar
This commit adds support for obtaining the current avatar for an user
with the botogram.User.avatar attribute.

Because executing a request for every User object when 99% of the times
you don't need the avatar really kills performances, the content of the
attribute is dynamically loaded the first time you access it.