Let's welcome botogram 0.2
botogram 0.2 is the second alpha release of botogram. It features an
increased support for the upstream Telegram API, and also some bugfixes
here and there.

This release also does some cleanup in the API, providing better methods
and deprecating the old ones. The deprecated methods will be available
until botogram 1.0, and warnings are in place to notify you where to
change what.