Use HTML instead of Markdown in the default messages
Before this commit, Markdown was used in all the default messages. This
was fine for most of them, but it causes problems in the `/help`
command if someone puts any markdown special characters in the
docstring of any commands.

This commit switches all the default messages to forced HTML, because
it supports escaping unwanted entities.

Fixes: GH-51