Refactor the functions calling process
User-provided functions are currently not called directly, but via a
special "call" function which provides arguments dynamically (you can
request just a subset of the available ones without no worries).

Before this commit, the implementation was directly in the FrozenBot's
code, and didn't support lazily loaded arguments. This commit rewrites
parts of the implementation, moves it to the `utils` module, adds more
unit tests and adds support for lazily loaded arguments.

Lazily loaded arguments are especially useful because some arguments
might be slow to be prepared, and preparing them when the function
isn't requesting them is just a waste.