Return the Message from every send* method
Before this commit, there was no way to access the message you sent to
an user, and that's needed in order to, for example, editing existing
messages. Since this commit, the instance of botogram.Message relative
to the message you sent is returned from the methods you use to
communicate with the users.

Issue: GH-55