Properly handle multiple instances of the same bot
Telegram doesn't allow multiple instances of the same bot to live
together, and previous this commit that behavior resulted in a bunch of
"Error fetching the updates" messages spammed to the bot logs.

This commit properly handles the situation, providing only a nice
message when the thing happens and waiting for the other instances to
shut down in the background. Then, after botogram is reasonably sure
the other instances are down, a message is displayed to the user and
the bot starts working again.

This definitely improves the situation in those cases.