Use the hidden=True argument on @bot.command to hide commands
Before this commit, the only way to hide existing commands from the
/help command was to manage the bot.hide_commands list, but the
approach had some pitfalls: it was easy to override previous entries,
and there was no way for components to insert commands in it.

This new commit deprecates that old behavior, and refactors the whole
commands hiding system: now it's possible to pass hidden=True to
@bot.command (main component) or self.add_command (custom components)
and have the command hidden from the message.

Fixes: GH-19