Add support for channels in Message.forward_from
A recent update to the Bot API introduced a new `forward_from_chat`
attribute on the Message object, which contains the channel in which
the forwarded message was sent originally.

The problem with that was, having both a `forward_from` and a
`forward_from_chat` attributes is at least confusing, and not a great
user experience. Because of that, the two attributes are now merged
into a single one.