Add Bot.chat and deprecate Bot.send*
The Bot API 2.1 update introduced a new API call to get all information
about a chat from its ID, which can be used to recreate a Chat object.

Because of that, all the send* methods on the Bot class were retundant,
since you can do `bot.chat(123).send("Message")`. Because of that, all
those methods are now deprecated, and they will be removed in botogram
1.0.

Fixes: GH-65