Fix botogram crashing if someone edits a message
The Bot API 2.1 update (not supported by botogram 0.2.x) introduced a
new kind of Update when someone edits a message the bot knows about.
This allows to create cool things, but it isn't supported by botogram
0.2.x, since support for it was added in botogram 0.3.1.

Unfortunately, this meant botogram crashed when it received that kind
of update. This commit ignores every update other than the ones it
supports, so the fix lasts even for new kind of updates not introduced
yet.

Fixes: GH-70