Catch InterruptedError while interacting with sockets in the IPC
When someone sends a signal to a process, every system call running in
it simply interrupts for reasons. Python 3.5 automatically retries the
call, but Python 3.4 raises an InterruptedError.

This commit makes sure if a signal is received on Python 3.4 and
reading/writing on sockets in the IPC server is interrupted, the call
is retried.